+++ DTCD-WorkspacePanel/src/Plugin.js
@@ -13,6 +13,7 @@ import {
 export class WorkspacePanel extends AppPanelPlugin {
   #vueComponent;
   #logSystem;
+  #eventSystem;
 
   #config = {
     elementSize: 'medium',
@@ -26,6 +27,10 @@ export class WorkspacePanel extends AppPanelPlugin {
     super();
 
     const eventSystem = new EventSystemAdapter('0.4.0', guid);
+    this.#eventSystem = eventSystem;
+
+    this.#eventSystem.registerPluginInstance(this, ['WorkspaceDeleted']);
+
     const interactionSystem = new InteractionSystemAdapter('0.4.0');
     const workspaceSystem = new WorkspaceSystemAdapter('0.4.0');
     const router = new RouteSystemAdapter('0.1.0');
@@ -143,10 +148,13 @@ export class WorkspacePanel extends AppPanelPlugin {
             handler: {
               event: 'click',
               callback: permissions.delete
-                ? () => {
+                ? async () => {
                   const elemType = selectedElement.is_dir ? 'папку' : 'дашборд';
                   const isDelete = confirm(`Удалить ${elemType} "${selectedElement.title}"?`);
-                  isDelete && this.#vueComponent.deleteElement(selectedElement);
+                  if (isDelete) {
+                    await this.#vueComponent.deleteElement(selectedElement);
+                    this.#eventSystem.publishEvent('WorkspaceDeleted', {guid: this.guid});
+                  }
                 }
                 : undefined,
             },