+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- publish of event of workspace delete
 - logging of interactions with the server (deleting, creation, exporting, etc.)
 - permissions creating, deleting, editing and reading elements
 
@@ -22,13 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ability to create folders
 - filtering and sorting desktop elements
-- logging of interactions with the server (deleting, creation, exporting, etc.)
-- permissions creating, deleting, editing and reading elements
 
 ### Changed
 
 - updated functionality of the element creation modal window
-- form settings for element
 
 ### Fixed